+++ README.md
@@ -30,7 +30,7 @@
 - [ ] 사용자 도메인 추가
   - [ ] 필드: 이름, 이메일, 비밀번호
 - [ ] 로그인 기능 구현
-  - [ ] 로그인 페이지 조회
+  - [x] 로그인 페이지 조회
   - [ ] 로그인 API
 - [ ] 사용자 정보 조회
   - [ ] 사용자 정보 조회 API
