--- src/main/java/roomescape/controller/StaticAdminPageController.java
@@ -1,29 +1,0 @@
-package roomescape.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/admin")
-public class StaticAdminPageController {
-    @GetMapping
-    public String getAdminPage() {
-        return "admin/index";
-    }
-
-    @GetMapping("/reservation")
-    public String getReservationPage() {
-        return "admin/reservation-new";
-    }
-
-    @GetMapping("/time")
-    public String getReservationTimePage() {
-        return "admin/time";
-    }
-
-    @GetMapping("/theme")
-    public String getThemePage() {
-        return "admin/theme";
-    }
-}
