--- src/main/java/roomescape/controller/StaticUserPageController.java
@@ -1,18 +1,0 @@
-package roomescape.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class StaticUserPageController {
-
-    @GetMapping
-    public String getHome() {
-        return "index";
-    }
-
-    @GetMapping("/reservation")
-    public String getReservation() {
-        return "reservation";
-    }
-}
